+++ src/math.jl
@@ -109,28 +109,31 @@ function result(val::Complex, der::Complex, a::Measurement)
     return complex(result(real(val), real(der), a), result(imag(val), imag(der), a))
 end
 
-# @uncertain macro.  TODO: generalize to any number of arguments.
+### @uncertain macro.
 """
-    @uncertain f(value ± stddev[, value ± stddev])
+    @uncertain f(value ± stddev, ...)
 
 A macro to calculate \$f(value) ± uncertainty\$, with \$uncertainty\$ derived
-from \$stddev\$ according to rules of linear error propagation theory.  Function
-\$f\$ can accept one or two real argument, the type of the arguments provided
-must be `Measurement`.
+from \$stddev\$ according to rules of linear error propagation theory.
+
+Function \$f\$ can accept any number of real arguments, the type of the
+arguments provided must be `Measurement`.
 """
 macro uncertain(expr::Expr)
     f = esc(expr.args[1]) # Function name
-    if length(expr.args) == 2
+    n = length(expr.args) - 1
+    if n == 1
         a = esc(expr.args[2]) # Argument, of Measurement type
-        return :( Measurements.result($f($a.val), Calculus.derivative($f, $a.val), $a) )
-    elseif length(expr.args) == 3
-        a1 = esc(expr.args[2]) # First argument
-        a2 = esc(expr.args[3]) # Second argument
-        return :( result($f($a1.val, $a2.val),
-                         (Calculus.gradient(x -> $f(x...), [$a1.val, $a2.val])...),
-                         ($a1, $a2)) )
+        return :( result($f($a.val), Calculus.derivative($f, $a.val), $a) )
     else
-        throw(ArgumentError("@uncertain only supports functions with one or two arguments"))
+        a = expr.args[2:end] # Arguments, as an array of expressions
+        args = :([])  # Build up array of arguments
+        [push!(args.args, :($(esc(a[i])))) for i=1:n] # Fill the array
+        argsval =:([])  # Build up the array of values of arguments
+        [push!(argsval.args, :($(args.args[i]).val)) for i=1:n] # Fill the array
+        return :( result($f($argsval...),
+                         (Calculus.gradient(x -> $f(x...), $argsval)...),
+                         ($args...)) )
     end
 end